Edit patient API - Names (#2860)
* Edit Patient API - Edit names

* Change PatientNameDemographic.identifier to sequence. Re-generate UI models

* Fix odd formatting

+++ libs/database-entities/src/main/java/gov/cdc/nbs/patient/PatientCommand.java
@@ -448,7 +448,7 @@ public sealed interface PatientCommand {
 
   record DeleteNameInfo(
       long person,
-      int sequence,
+      short sequence,
       long requester,
       LocalDateTime requestedOn
   ) implements PatientCommand {
